Add grooming centres module

+++ db/migrate/20171205120622_create_clinics.rb
@@ -5,7 +5,7 @@ class CreateClinics < ActiveRecord::Migration[5.1]
       t.string :email
       t.string :picture
       t.string :mobile_number
-      t.float :consultation_fee
+      t.integer :consultation_fee
       t.string :website
       t.text :description
       t.boolean :is_active, default: false
